Refactor messages.
