+++ client/src/components/ManageAccount/tests/ManageAccount.test.js
@@ -8,7 +8,7 @@ import { initialState as profile } from '../../../redux/reducers/profile.reducer
 import { initialState as workbench } from '../../../redux/reducers/workbench.reducers';
 import { initialState as notifications } from '../../../redux/reducers/notifications.reducers';
 import { initialTicketCreateState as ticketCreate } from '../../../redux/reducers/tickets.reducers';
-import introMessages from '../../../redux/reducers/intro.reducers';
+import introMessages from '../../../redux/reducers/message.reducers';
 import ManageAccountPage from '../index';
 
 const mockStore = configureStore();
