--- server/portal/apps/intromessages/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class IntromessagesConfig(AppConfig):
-    name = 'intromessages'
-    app_label = 'intromessages'
