--- server/portal/apps/intromessages/urls.py
@@ -1,11 +1,0 @@
-"""IntroMessages URLs
-"""
-from django.urls import path
-from portal.apps.intromessages import views
-
-
-app_name = 'intromessages'
-urlpatterns = [
-    path('', views.IntroMessagesView.as_view()),
-    path('custom/', views.CustomMessagesView.as_view())
-]
