+++ server/portal/apps/message/models.py
@@ -46,7 +46,7 @@ class CustomMessageTemplate(models.Model):
     component = models.CharField(help_text='Component type', max_length=20, choices=COMPONENTS, default='Dashboard')
     message_type = models.CharField(help_text='Message type', max_length=20, choices=MESSAGE_TYPES, default='info')
     dismissible = models.BooleanField(default=True)
-    message = models.CharField(help_text='Message content (max 200 chars)', max_length=200, default='', blank=True)
+    message = models.CharField(help_text='Message content (max 200 characters)', max_length=200, default='', blank=True)
 
     def to_dict(self):
         return {
