+++ server/portal/urls.py
@@ -77,8 +77,8 @@ urlpatterns = [
     path('request-access/', include('portal.apps.request_access.urls', namespace='request_access')),
     path('search/', include('portal.apps.site_search.urls', namespace='site_search')),
 
-    # intromessages
-    path('api/intromessages/', include('portal.apps.intromessages.urls', namespace='intromessages')),
+    # message
+    path('api/message/', include('portal.apps.message.urls', namespace='message')),
 
 
     # integrations
